feat(api): use a dedicated admin-combined-course-blueprint-serializer
