+++ api/src/quest/infrastructure/serializers/combined-course-blueprint-serializer.js
@@ -1,29 +1,11 @@
 import jsonapiSerializer from 'jsonapi-serializer';
 
-import { AdminCombinedCourseBlueprint } from '../../domain/models/AdminCombinedCourseBlueprint.js';
-
-const { Deserializer, Serializer } = jsonapiSerializer;
+const { Serializer } = jsonapiSerializer;
 
 const serialize = function (combinedCourseBlueprint) {
   return new Serializer('combined-course-blueprints', {
-    attributes: [
-      'name',
-      'internalName',
-      'description',
-      'illustration',
-      'content',
-      'createdAt',
-      'updatedAt',
-      'attestationKey',
-    ],
+    attributes: ['name', 'internalName', 'description', 'illustration', 'createdAt', 'updatedAt'],
   }).serialize(combinedCourseBlueprint);
 };
 
-const deserialize = async function (payload) {
-  const deserializedData = await new Deserializer({
-    keyForAttribute: 'camelCase',
-  }).deserialize(payload);
-  return new AdminCombinedCourseBlueprint(deserializedData);
-};
-
-export { deserialize, serialize };
+export { serialize };
